Separated old knowledge processor commands
